requirements.txt
tda-api
pandas
arrow